+++ frontend-project/src/pages/cases/list/index.tsx
@@ -1,12 +1,11 @@
-import { PageHeaderWrapper } from '@ant-design/pro-layout';
-import ProTable, { ProColumns, ActionType } from '@ant-design/pro-table';
-import React, { FC, useRef } from 'react';
+import { ProColumns } from '@ant-design/pro-table';
+import React, { FC } from 'react';
 import { formatMessage } from 'umi-plugin-react/locale';
 
 import { Case, fetchCasesPage } from '@/services/cases';
+import Table from '@/components/Table';
 
 const TableList: FC<{}> = () => {
-  const actionRef = useRef<ActionType>();
   const columns: ProColumns<Case>[] = [
     {
       title: formatMessage({ id: 'cases-list.table.columns.name.title' }),
@@ -33,30 +32,7 @@ const TableList: FC<{}> = () => {
       dataIndex: 'tags',
     },
   ];
-
-  const showTotal = (total, range) =>
-    `${range[0]}-${range[1]} / ${formatMessage({ id: 'cases-list.table.total' })} ${total}`;
-
-  return (
-    <PageHeaderWrapper content={formatMessage({ id: 'cases-list.page-header-content' })}>
-      <ProTable<Case>
-        headerTitle={formatMessage({ id: 'cases-list.table-header-title' })}
-        actionRef={actionRef}
-        rowKey="id"
-        tableAlertRender={false}
-        request={fetchCasesPage}
-        columns={columns}
-        rowSelection={false}
-        search={false}
-        options={false}
-        pagination={{
-          pageSize: 20,
-          showSizeChanger: false,
-          showTotal,
-        }}
-      />
-    </PageHeaderWrapper>
-  );
+  return <Table type="cases" columns={columns} fetchData={fetchCasesPage} />;
 };
 
 export default TableList;
