+++ frontend-project/src/services/cases.ts
@@ -26,13 +26,7 @@ export async function fetchCasesPage({
     limit: pageSize,
     offset: pageSize * (current - 1),
   });
-  sdkResponse.results = sdkResponse.results.map(item => ({
-    ...item,
-    createdOn: item.createdOn.toLocaleString(),
-    modifiedOn: item.modifiedOn.toLocaleString(),
-  }));
 
-  console.log(sdkResponse.results[0].createdOn);
   return {
     data: sdkResponse.results,
     total: sdkResponse.count,
